# Get rid of unnecessary reinitialization of the `makeShareableClone` function