--- a/apps/common-app/src/apps/reanimated/examples/CopyShareablesPerformanceTest.tsx
+++ b/apps/common-app/src/apps/reanimated/examples/CopyShareablesPerformanceTest.tsx
@@ -1,5 +1,5 @@
-import { View, Text, Button, TextInput, StyleSheet, TouchableWithoutFeedback, Keyboard } from 'react-native';
 import { useState } from 'react';
+import { Button, Keyboard,StyleSheet, Text, TextInput, TouchableWithoutFeedback, View } from 'react-native';
 import { makeShareableCloneRecursive } from 'react-native-reanimated';
 
 function createRandomObject(numberOfKeys: number) {