## Summary
Get rid of unnecessary reinitialization of the `makeShareableClone` function.

## Test plan
From tests I've done, it seems to speed up copying large objects.
| Number of Objects | Number of Keys | Before [ms] | After [ms] |
|-------------------|----------------|-------------|------------|
| 100               | 100            | 54.22       | 28.07      |
| 100               | 200            | 96.49       | 77.97      |
| 100               | 500            | 213.49      | 151.89     |
| 100               | 1000           | 570.34      | 286.37     |
| 200               | 200            | 165.17      | 126.58     |
| 200               | 500            | 384.98      | 291.73     |
| 200               | 1000           | 735.61      | 547.24     |
| 500               | 500            | 977.62      | 688.93     |
| 500               | 1000           | 1922.90     | 1418.57    |
| 1000              | 1000            | 3762.55     | 2807.86    |
<video src="https://github.com/user-attachments/assets/cb0678ee-c572-4f7a-a7be-4ea2895c0e9f" />